Mobile fix; header image fix
Added "!important;" to line 37 of mobile.scss to force the "box-2" element to comply with 100%-20px request.

Changed name of "big-header.jpeg" to "header-bg.jpeg" to simplify re-adding the image when the theme component breaks due to update. Simply re-upload the image file, instead of re-upload and rename each time.